+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v6.5.60
+----------
+ * Add option to audit_archives to check flow run counts
+ * Drop no longer used ticket subject column
+ * Add contact read page based on contact chat component
+
 v6.5.59
 ----------
  * Less progress updates in audit_archives